enable all compiler warnings and fixed them -- mostly signed vs unsigned compare

+++ include/contigs.hpp
@@ -136,14 +136,14 @@ public:
     return contigs.end();
   }
 
-  void print_stats(int min_ctg_len) {
+  void print_stats(unsigned min_ctg_len) {
     BarrierTimer timer(__FILEFUNC__);
     int64_t tot_len = 0, max_len = 0;
     double tot_depth = 0;
-    vector<pair<int, int64_t>> length_sums({ {1, 0}, {5, 0}, {10, 0}, {25, 0}, {50, 0}});
+    vector<pair<unsigned, uint64_t>> length_sums({ {1, 0}, {5, 0}, {10, 0}, {25, 0}, {50, 0}});
     int64_t num_ctgs = 0;
     int64_t num_ns = 0;
-    vector<int> lens;
+    vector<unsigned> lens;
     lens.reserve(contigs.size());
     for (auto ctg : contigs) {
       auto len = ctg.seq.length();
@@ -210,7 +210,7 @@ public:
     }
   }
 
-  void dump_contigs(const string &fname, int min_ctg_len) {
+  void dump_contigs(const string &fname, unsigned min_ctg_len) {
     BarrierTimer timer(__FILEFUNC__);
     string fasta = "";
     for (auto it = contigs.begin(); it != contigs.end(); ++it) {
@@ -260,7 +260,7 @@ public:
     // these can be equal if the contigs are very long and there are many ranks so this one doesn't get even a full contig
     ctgs_file.seekg(start_offset);
     while (!ctgs_file.eof()) {
-      if (ctgs_file.tellg() >= stop_offset) break;
+      if ((size_t) ctgs_file.tellg() >= stop_offset) break;
       getline(ctgs_file, cname);
       if (cname == "") break;
       getline(ctgs_file, seq);
